+++ banyand/metadata/schema/etcd.go
@@ -19,6 +19,7 @@ package schema
 
 import (
 	"context"
+	"fmt"
 	"sync"
 	"time"
 
@@ -29,8 +30,6 @@ import (
 	"google.golang.org/protobuf/proto"
 
 	commonv1 "github.com/apache/skywalking-banyandb/api/proto/banyandb/common/v1"
-	databasev1 "github.com/apache/skywalking-banyandb/api/proto/banyandb/database/v1"
-	propertyv1 "github.com/apache/skywalking-banyandb/api/proto/banyandb/property/v1"
 	"github.com/apache/skywalking-banyandb/pkg/logger"
 	"github.com/apache/skywalking-banyandb/pkg/run"
 )
@@ -63,19 +62,11 @@ func ConfigureServerEndpoints(url []string) RegistryOption {
 	}
 }
 
-type eventHandler struct {
-	handler      EventHandler
-	interestKeys Kind
-}
-
-func (eh *eventHandler) interestOf(kind Kind) bool {
-	return KindMask&kind&eh.interestKeys != 0
-}
-
 type etcdSchemaRegistry struct {
 	client   *clientv3.Client
 	closer   *run.Closer
-	handlers []*eventHandler
+	l        *logger.Logger
+	watchers []*watcher
 	mux      sync.RWMutex
 }
 
@@ -83,33 +74,19 @@ type etcdSchemaRegistryConfig struct {
 	serverEndpoints []string
 }
 
-func (e *etcdSchemaRegistry) RegisterHandler(kind Kind, handler EventHandler) {
+func (e *etcdSchemaRegistry) RegisterHandler(name string, kind Kind, handler EventHandler) {
+	// Validate kind
+	if kind&KindMask != kind {
+		panic(fmt.Sprintf("invalid kind %d", kind))
+	}
 	e.mux.Lock()
 	defer e.mux.Unlock()
-	e.handlers = append(e.handlers, &eventHandler{
-		interestKeys: kind,
-		handler:      handler,
-	})
-}
-
-func (e *etcdSchemaRegistry) notifyUpdate(metadata Metadata) {
-	e.mux.RLock()
-	hh := e.handlers
-	e.mux.RUnlock()
-	for _, h := range hh {
-		if h.interestOf(metadata.Kind) {
-			h.handler.OnAddOrUpdate(metadata)
-		}
-	}
-}
-
-func (e *etcdSchemaRegistry) notifyDelete(metadata Metadata) {
-	e.mux.RLock()
-	hh := e.handlers
-	e.mux.RUnlock()
-	for _, h := range hh {
-		if h.interestOf(metadata.Kind) {
-			h.handler.OnDelete(metadata)
+	for i := 0; i < KindSize; i++ {
+		ki := Kind(1 << i)
+		if kind&ki > 0 {
+			e.l.Info().Str("name", name).Stringer("kind", kind).Msg("registering watcher")
+			w := e.newWatcher(name, ki, handler)
+			e.watchers = append(e.watchers, w)
 		}
 	}
 }
@@ -117,6 +94,10 @@ func (e *etcdSchemaRegistry) notifyDelete(metadata Metadata) {
 func (e *etcdSchemaRegistry) Close() error {
 	e.closer.Done()
 	e.closer.CloseThenWait()
+	for i, w := range e.watchers {
+		_ = w.Close()
+		e.watchers[i] = nil
+	}
 	return e.client.Close()
 }
 
@@ -129,7 +110,8 @@ func NewEtcdSchemaRegistry(options ...RegistryOption) (Registry, error) {
 	if registryConfig.serverEndpoints == nil {
 		return nil, errors.New("server address is not set")
 	}
-	zapCfg := logger.GetLogger("etcd").ToZapConfig()
+	log := logger.GetLogger("etcd")
+	zapCfg := log.ToZapConfig()
 
 	var l *zap.Logger
 	var err error
@@ -152,6 +134,7 @@ func NewEtcdSchemaRegistry(options ...RegistryOption) (Registry, error) {
 	reg := &etcdSchemaRegistry{
 		client: client,
 		closer: run.NewCloser(1),
+		l:      log,
 	}
 	return reg, nil
 }
@@ -207,7 +190,7 @@ func (e *etcdSchemaRegistry) update(ctx context.Context, metadata Metadata) erro
 	}
 	replace := getResp.Count > 0
 	if replace {
-		existingVal, innerErr := metadata.Unmarshal(getResp.Kvs[0].Value)
+		existingVal, innerErr := metadata.Kind.Unmarshal(getResp.Kvs[0])
 		if innerErr != nil {
 			return innerErr
 		}
@@ -230,7 +213,6 @@ func (e *etcdSchemaRegistry) update(ctx context.Context, metadata Metadata) erro
 	} else {
 		return ErrGRPCResourceNotFound
 	}
-	e.notifyUpdate(metadata)
 	return nil
 }
 
@@ -265,38 +247,27 @@ func (e *etcdSchemaRegistry) create(ctx context.Context, metadata Metadata) erro
 	if err != nil {
 		return err
 	}
-
-	e.notifyUpdate(metadata)
 	return nil
 }
 
-func (e *etcdSchemaRegistry) listWithPrefix(ctx context.Context, prefix string, factory func() proto.Message) ([]proto.Message, error) {
+func (e *etcdSchemaRegistry) listWithPrefix(ctx context.Context, prefix string, kind Kind) ([]proto.Message, error) {
 	if !e.closer.AddRunning() {
 		return nil, ErrClosed
 	}
 	defer e.closer.Done()
-	resp, err := e.client.Get(ctx, prefix, clientv3.WithFromKey(), clientv3.WithRange(incrementLastByte(prefix)))
+	resp, err := e.client.Get(ctx, prefix, clientv3.WithPrefix())
 	if err != nil {
 		return nil, err
 	}
 	entities := make([]proto.Message, resp.Count)
 	for i := int64(0); i < resp.Count; i++ {
-		message := factory()
-		if innerErr := proto.Unmarshal(resp.Kvs[i].Value, message); innerErr != nil {
-			return nil, innerErr
+		md, err := kind.Unmarshal(resp.Kvs[i])
+		if err != nil {
+			return nil, err
 		}
-		entities[i] = message
-		if messageWithMetadata, ok := message.(HasMetadata); ok {
-			// Assign readonly fields
-			messageWithMetadata.GetMetadata().CreateRevision = resp.Kvs[i].CreateRevision
-			messageWithMetadata.GetMetadata().ModRevision = resp.Kvs[i].ModRevision
-		}
+		entities[i] = md.Spec.(proto.Message)
 	}
 	return entities, nil
-}
-
-func listPrefixesForEntity(group, entityPrefix string) string {
-	return groupsKeyPrefix + group + entityPrefix
 }
 
 func (e *etcdSchemaRegistry) delete(ctx context.Context, metadata Metadata) (bool, error) {
@@ -313,33 +284,6 @@ func (e *etcdSchemaRegistry) delete(ctx context.Context, metadata Metadata) (boo
 		return false, err
 	}
 	if resp.Deleted == 1 {
-		var message proto.Message
-		switch metadata.Kind {
-		case KindMeasure:
-			message = &databasev1.Measure{}
-		case KindStream:
-			message = &databasev1.Stream{}
-		case KindIndexRuleBinding:
-			message = &databasev1.IndexRuleBinding{}
-		case KindIndexRule:
-			message = &databasev1.IndexRule{}
-		case KindProperty:
-			message = &propertyv1.Property{}
-		case KindTopNAggregation:
-			message = &databasev1.TopNAggregation{}
-		default:
-			return false, nil
-		}
-		if unmarshalErr := proto.Unmarshal(resp.PrevKvs[0].Value, message); unmarshalErr == nil {
-			e.notifyDelete(Metadata{
-				TypeMeta: TypeMeta{
-					Kind:  metadata.Kind,
-					Name:  metadata.Name,
-					Group: metadata.Group,
-				},
-				Spec: message,
-			})
-		}
 		return true, nil
 	}
 	return false, nil
@@ -404,8 +348,16 @@ func (e *etcdSchemaRegistry) register(ctx context.Context, metadata Metadata) er
 	return nil
 }
 
+func (e *etcdSchemaRegistry) newWatcher(name string, kind Kind, handler EventHandler) *watcher {
+	return newWatcher(e.client, kind, handler, e.l.Named(fmt.Sprintf("watcher-%s[%s]", name, kind.String())))
+}
+
+func listPrefixesForEntity(group, entityPrefix string) string {
+	return entityPrefix + "/" + group
+}
+
 func formatKey(entityPrefix string, metadata *commonv1.Metadata) string {
-	return groupsKeyPrefix + metadata.GetGroup() + entityPrefix + metadata.GetName()
+	return listPrefixesForEntity(metadata.GetGroup(), entityPrefix) + "/" + metadata.GetName()
 }
 
 func incrementLastByte(key string) string {